Create README.md
The project is an example of a simple todo app, created using ReactJS, HTML, and CSS. On the Home page, new tasks can be added to the app using the input field and add button, and the task can be selected as active or inactive depending on the situation. The user can also delete the task completely if desired. In addition, a navbar menu, about us, and contact pages have been added to the app. The project is not connected to any database and has been created statically.

Follow the steps to set up the project:

Go to the official website of Node.js and download the latest version.
Run the downloaded file to complete the installation.
After the installation is complete, open a terminal or command prompt and run the node -v command to check if Node.js has been successfully installed.


In the same terminal or command prompt, run the npm -v command to check if npm has been successfully installed.
First, go to the GitHub page of the project and click on the “Code” button in the top right corner. Then, select “Download ZIP” to download the project to your computer.
Unzip the downloaded ZIP file and go to the resulting folder.
Open a terminal or command prompt and navigate to the folder where the project is located.
Run the npm install command to install the project's dependencies.
Run the npm start command to start the project.
After following these steps, your project will start running on your local server. You can view the project by going to http://localhost:3000 in your browser. You can find images related to the project below.
NOTE:The project is only compatible with desktop web in terms .